+++ lib/app/model/pojo/response/ProductRespone.dart
@@ -48,7 +48,7 @@ class ProductData {
   List<ProductImage> image;
   int discountPercent;
   int rating;
-  int reviewCount;
+  double reviewCount;
   int stockQuantity;
 
   ProductData(
